FIX: Make pd import local in DataFrameTransformer (#1017)
DataFrameTransformer requires pandas but we don't want to have a
dependency on pandas. Therefore, the pandas import should be local.

Back when I wrote this, I thought it would be sufficient to import
pandas at a class level, but that is incorrect. Now, pandas is important
inside the method bodies.

For most skorch users, this should make no difference. But it will
allow skorch users who do use something from helpers.py to avoid having
to install pandas.