Release 1.0.0 (#1055)
* Update version

* Update CHANGES.md

* Remove test skip for old PyTorch versions

Only needed for PyTorch 1.11 and lower, which are no longer supported in
the CI.

Release message:

The 1.0.0 release of skorch is here. We think that skorch is at a very
stable point, which is why a 1.0.0 release is appropriate. There are no
plans to add any breaking changes or major revisions in the future.
Instead, our focus now is to keep skorch up-to-date with the latest
versions of PyTorch and scikit-learn, and to fix any bugs that may
arise.