ENH ReduceLROnPlateau record lr, works on batches
Previously, when using ReduceLROnPlateau, we would not record the
learning rates in history. The comment says that's because this class
does not expose the get_last_lr method. I checked it again and it's now
present, so let's use it.

Furthermore, I made a change to enable ReduceLROnPlateau to step on each
batch instead of each epoch. This is consistent with other learning rate
schedulers.