FIX Make skorch work with sklearn 1.5
The new sklearn version now requires the estimator to expose a method
__sklearn_tags__. This PR implements a bare minimum of sklearn tags to
make the tests pass again.

Probably this list of tags can be improved to be more precise.

Moreover, some tests now started failing. To be precise, using
GridSearchCV with y being a torch tensor fails, as sklearn performs a
check on the devices of y vs y_pred (a numpy array) and determines that
the devices differ. The error message is:

    def device(*array_list, remove_none=True, remove_types=(str,)):
        """Hardware device where the array data resides on.

        If the hardware device is not the same for all arrays, an error is raised.

        Parameters
        ----------
        *array_list : arrays
            List of array instances from NumPy or an array API compatible library.

        remove_none : bool, default=True
            Whether to ignore None objects passed in array_list.

        remove_types : tuple or list, default=(str,)
            Types to ignore in array_list.

        Returns
        -------
        out : device
            `device` object (see the "Device Support" section of the array API spec).
        """
        array_list = _remove_non_arrays(
            *array_list, remove_none=remove_none, remove_types=remove_types
        )

        if not array_list:
            return None

        device_ = _single_array_device(array_list[0])

        # Note: here we cannot simply use a Python `set` as it requires
        # hashable members which is not guaranteed for Array API device
        # objects. In particular, CuPy devices are not hashable at the
        # time of writing.
        for array in array_list[1:]:
            device_other = _single_array_device(array)
            if device_ != device_other:
>               raise ValueError(
                    f"Input arrays use different devices: {str(device_)}, "
                    f"{str(device_other)}"
                )
E               ValueError: Input arrays use different devices: cpu, cpu

The tests have now been amended to cast y to a numpy array. It is
unfortunate that this used not be necessary but now is.

I think we can live with that, as it's an edge case. I tested that with
normal fitting, passing a torch.tensor as y is no problem.