Add step_args parameter to LRScheduler.simulate
The purpose of this change is to allow for simulation of scheduling policies
such as `ReduceLROnPlateau` to be simulated properly.

If `step_args` is an indexable object, it is indexed using the current
simulated epoch to get closer to real-life behavior, e.g. simulating a
real loss curve and the corresponding behavior of the LR scheduler policy.