Unrelated: Fix issues with optional pandas dep
For a pandas-related fixture, we had a pytest.mark.skipif decorator.
Apparently, that doesn't work and the new pytest now raises an error
because of that. Therefore, the decorator is now moved to the class
level.

While testing this, I noticed that if pandas is not installed, trying to
import sacred results in an AttributeError. Therefore, the check whether
sacred is installed is now checking for pandas first.