FIX Compatibility with sklearn v1.8
There were two issues in skorch with sklearn v1.8 that are now fixed:

__sklearn_is_fitted__ should return a bool. Instead, we were just
raising or returning None. Not sure if this was already wrong when
added (#1119) or if the API changed later.

Moreover, sklearn was now raising an error when using the SkorchDoctor
class because of missing __sklearn_tags__. This class now inherits from
BaseEstimator, resolving this issue.

For good measure, I also tested locally with sklearn 1.7 to ensure that
these changes are backwards compatible.

Unrelated changes:

Remove an unnecessary import and trailing whitespace. Added
__sklearn_is_fitted__ to SkorchDoctor too.